CPLP-244: move model classes to service-module
