+++ coreservices/useradministration/src/CatenaX.NetworkServices.UserAdministration.Service/Models/UserUpdateBpn.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-namespace CatenaX.NetworkServices.Provisioning.Library
+namespace CatenaX.NetworkServices.UserAdministration.Service.Models
 {
     public class UserUpdateBpn
     {
