+++ coreservices/useradministration/src/CatenaX.NetworkServices.UserAdministration.Service/Models/UserUpdateInfo.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
-namespace CatenaX.NetworkServices.Provisioning.Library
+namespace CatenaX.NetworkServices.UserAdministration.Service.Models
+
 {
     public class UserIds
     {
